refactor for docker setup
