+++ lavalink/application.yml
